RangeValueConverter: handle bindable error on Jumplist.SelectedItem when using Unicode ranges in JumpList. Ranges: - Add Find(string) for RangeValueConverter Range: - Remove unnecessry casting.
FontGlypsViewModel:
- Clear Search.ShowResults when the jumplist is open

GlyphsView:
- Remove unnecessary #region/#endregion
SearchViewModel:
- clear SearchText when metric properties change.

FontGlyphsView:
- Use RangeValueConverter for Jumplist SelectedItem binding.

+++ Text/Unicode/Range.cs
@@ -94,7 +94,7 @@ public readonly struct Range : IEquatable<Range>
         }
         First = first;
         Last = last;
-        Length = ((uint)last - (uint)first) + 1;
+        Length = last - first + 1;
         Name = name;
     }
